+++ src/download.cpp
@@ -250,8 +250,8 @@ static int lzrc_decompress(void* out, int out_len, const void* in, int in_len)
             if (rc.out_ptr == rc.out_len)
             {
                 throw DownloadError(
-                        "internal error - lzrc output overflow! pkg may be "
-                        "corrupted");
+                        "内部错误 - PKG文件不完整或已损坏 ! 请重新"
+                        "下载");
             }
             rc.output[rc.out_ptr++] = (uint8_t)byte;
             last_byte = (uint8_t)byte;
@@ -318,15 +318,15 @@ static int lzrc_decompress(void* out, int out_len, const void* in, int in_len)
             if (match_dist > rc.out_ptr)
             {
                 throw DownloadError(
-                        "internal error - lzrc match_dist out of range! pkg "
-                        "may be corrupted");
+                        "内部错误 - PKG文件不完整或已损坏! "
+                        "请重新下载");
             }
 
             if (rc.out_ptr + match_len + 1 > rc.out_len)
             {
                 throw DownloadError(
-                        "internal error - lzrc output overflow! pkg may be "
-                        "corrupted");
+                        "内部错误 - PKG文件不完整或已损坏! 请重新"
+                        "下载");
             }
 
             const uint8_t* match_src = rc.output + rc.out_ptr - match_dist;
@@ -383,7 +383,7 @@ void Download::download_data(
         uint8_t* buffer, uint32_t size, int encrypted, int save)
 {
     if (is_canceled())
-        throw std::runtime_error("download was canceled");
+        throw std::runtime_error("下载已被取消");
 
     if (size == 0)
         return;
@@ -398,7 +398,7 @@ void Download::download_data(
         const int64_t http_length = _http->get_length();
         if (http_length < 0)
         {
-            throw DownloadError("HTTP response has unknown length");
+            throw DownloadError("HTTP响应长度未知");
         }
 
         download_size = http_length + download_offset;
@@ -416,7 +416,7 @@ void Download::download_data(
         {
             const int read = _http->read(buffer + pos, size - pos);
             if (read == 0)
-                throw DownloadError("HTTP connection closed");
+                throw DownloadError("HTTP连接意外断开");
             pos += read;
         }
     }
@@ -445,7 +445,7 @@ void Download::download_data(
         }
 
         if (!pkgi_write(item_file, buffer, write))
-            throw formatEx<DownloadError>("failed to write to {}", item_path);
+            throw formatEx<DownloadError>("写入至 {} 失败", item_path);
     }
 }
 
@@ -453,7 +453,7 @@ void Download::skip_to_file_offset(uint64_t to_offset)
 {
     if (to_offset < encrypted_offset)
         throw DownloadError(
-                fmt::format("can't seek backward to {}", to_offset));
+                fmt::format("无法向后寻找至 {}", to_offset));
 
     std::vector<uint8_t> down(64 * 1024);
     while (encrypted_offset != to_offset)
@@ -481,7 +481,7 @@ void Download::create_file()
     LOGF("creating {} file", item_name);
     item_file = pkgi_create(item_path.c_str());
     if (!item_file)
-        throw formatEx<DownloadError>("cannot create file {}", item_name);
+        throw formatEx<DownloadError>("无法创建 {} 文件", item_name);
 }
 
 void Download::open_file()
@@ -489,7 +489,7 @@ void Download::open_file()
     LOGF("opening {} file for resume", item_name);
     item_file = pkgi_openrw(item_path.c_str());
     if (!item_file)
-        throw formatEx<DownloadError>("cannot create file {}", item_name);
+        throw formatEx<DownloadError>("无法创建 {} 文件", item_name);
 }
 
 int Download::download_head(const uint8_t* rif)
@@ -516,14 +516,14 @@ int Download::download_head(const uint8_t* rif)
     if (get32be(head.data()) != 0x7f504b47 ||
         get32be(head.data() + PKG_HEADER_SIZE) != 0x7F657874)
     {
-        throw DownloadError("wrong pkg header");
+        throw DownloadError("错误的PKG文件头");
     }
 
     // contentid is at 0x50 for psm games
     if (rif && !(pkgi_memequ(rif + 0x10, head.data() + 0x30, 0x30) ||
                  pkgi_memequ(rif + 0x50, head.data() + 0x30, 0x30)))
     {
-        throw DownloadError("zRIF content id doesn't match pkg");
+        throw DownloadError("zRIF所包含的数据与PKG文件不匹配");
     }
 
     const auto meta_offset = get32be(head.data() + 8);
@@ -568,7 +568,7 @@ int Download::download_head(const uint8_t* rif)
         aes128_encrypt(&ctx, iv, key);
     }
     else
-        throw DownloadError("invalid key type " + std::to_string(key_type));
+        throw DownloadError("无效的密钥类型" + std::to_string(key_type));
 
     aes128_ctr_init(&aes, key);
 
@@ -586,7 +586,7 @@ int Download::download_head(const uint8_t* rif)
     {
         if (offset + 16 >= enc_offset)
         {
-            throw DownloadError("pkg file is too small or corrupted");
+            throw DownloadError("PKG文件不完整或已损坏");
         }
 
         uint32_t type = get32be(head.data() + offset + 0);
@@ -606,7 +606,7 @@ int Download::download_head(const uint8_t* rif)
                 content_type != CONTENT_TYPE_PSV_DLC)
             {
                 throw DownloadError(
-                        "unsupported package type: " +
+                        "不支持的PKG类型: " +
                         std::to_string(content_type));
             }
         }
@@ -633,9 +633,9 @@ int Download::download_head(const uint8_t* rif)
     if (index_size && item_offset != index_size)
     {
         throw DownloadError(
-                "assertion error: read-ahead mismatch, expected: " +
+                "声明错误: 预读与期望不匹配: " +
                 std::to_string(index_size) +
-                ", got: " + std::to_string(item_offset));
+                ", 实际: " + std::to_string(item_offset));
     }
 
     const auto target_size = (uint32_t)(enc_offset + item_offset);
@@ -669,19 +669,19 @@ void Download::download_file_content(uint64_t encrypted_size)
 void Download::download_file_content_to_iso(uint64_t item_size)
 {
     if (item_size < 0x28)
-        throw DownloadError("eboot.pbp file is too small");
+        throw DownloadError("eboot.pbp文件错误");
 
     uint8_t eboot_header[0x28];
     download_data(eboot_header, sizeof(eboot_header), 1, 0);
 
     if (memcmp(eboot_header, "\x00PBP", 4) != 0)
-        throw DownloadError("wrong eboot.pbp header magic");
+        throw DownloadError("错误的eboot.pbp文件头");
 
     uint32_t const psar_offset = get32le(eboot_header + 0x24);
     if (psar_offset + 256 > item_size)
-        throw DownloadError("eboot.pbp file is to short");
+        throw DownloadError("eboot.pbp文件错误");
     if (psar_offset % 16 != 0)
-        throw DownloadError("psar_offset is not aligned");
+        throw DownloadError("psar_offset偏移量未对齐");
 
     skip_to_file_offset(psar_offset);
 
@@ -689,12 +689,12 @@ void Download::download_file_content_to_iso(uint64_t item_size)
     download_data(psar_header.data(), psar_header.size(), 1, 0);
 
     if (memcmp(psar_header.data(), "NPUMDIMG", 8) != 0)
-        throw DownloadError("wrong data.psar header magic");
+        throw DownloadError("错误的data.psar文件头");
 
     uint32_t const iso_block = get32le(psar_header.data() + 0x0c);
     if (iso_block > 16)
         throw DownloadError(fmt::format(
-                "unsupported data.psar block size %u, max %u supported",
+                "不支持的data.psar数据大小 %u, 最大支持 %u",
                 iso_block,
                 16));
 
@@ -714,7 +714,7 @@ void Download::download_file_content_to_iso(uint64_t item_size)
     uint32_t iso_table = get32le(psar_header.data() + 0x6c);
 
     if (iso_table + block_count * 32 > item_size)
-        throw DownloadError("offset table in data.psar file is too large");
+        throw DownloadError("data.psar文件中的偏移量表过大");
 
     uint64_t const table_offset = psar_offset + iso_table;
     skip_to_file_offset(table_offset);
@@ -737,7 +737,7 @@ void Download::download_file_content_to_iso(uint64_t item_size)
 
         if (psar_offset + block_size > item_size)
             throw DownloadError(fmt::format(
-                    "iso block size/offset is to large: {} > {}",
+                    "ISO数据块大小/偏移量过大: {} > {}",
                     psar_offset + block_size,
                     item_size));
 
@@ -761,7 +761,7 @@ void Download::download_file_content_to_iso(uint64_t item_size)
         {
             if (!pkgi_write(item_file, data.data(), block_size))
                 throw DownloadError(
-                        fmt::format("failed to write to %s", item_path));
+                        fmt::format("无法写入至 %s", item_path));
         }
         else
         {
@@ -774,12 +774,12 @@ void Download::download_file_content_to_iso(uint64_t item_size)
             if (out_size != int(iso_block) * ISO_SECTOR_SIZE)
             {
                 throw DownloadError(
-                        "internal error - lzrc decompression failed! "
-                        "pkg may be corrupted");
+                        "内部错误 - PKG文件可能已损坏! "
+                        "请重新下载");
             }
             if (!pkgi_write(item_file, uncompressed.data(), out_size))
                 throw DownloadError(
-                        fmt::format("failed to write to %s", item_path));
+                        fmt::format("无法写入至 %s", item_path));
         }
     }
 
@@ -789,7 +789,7 @@ void Download::download_file_content_to_iso(uint64_t item_size)
 void Download::download_file_content_to_edat(uint64_t item_size)
 {
     if (item_size < 0x90 + 0xa0)
-        throw DownloadError("EDAT file is too short");
+        throw DownloadError("EDAT文件过小");
 
     uint8_t item_header[0x5a];
     download_data(item_header, sizeof(item_header), 1, 0);
@@ -801,13 +801,12 @@ void Download::download_file_content_to_edat(uint64_t item_size)
     download_data(key_header, sizeof(key_header), 1, 0);
 
     if (memcmp(key_header, "\x00PGD", 4) != 0)
-        throw DownloadError("wrong EDAT header magic");
+        throw DownloadError("错误的EDAT文件头");
 
     uint32_t key_index = get32le(key_header + 4);
     uint32_t drm_type = get32le(key_header + 8);
-
     if (key_index != 1 || drm_type != 1)
-        throw DownloadError("unsupported EDAT file, key/drm type is wrong");
+        throw DownloadError("不支持的EDAT文件, key/drm类型错误");
 
     uint8_t mac[16];
     aes128_cmac(kirk7_key38, key_header, 0x70, mac);
@@ -821,10 +820,9 @@ void Download::download_file_content_to_edat(uint64_t item_size)
     uint32_t data_offset = get32le(key_header + 0x4c);
 
     if (data_offset != 0x90)
-        throw DownloadError("unsupported EDAT file, data/offset is wrong");
+        throw DownloadError("不支持的EDAT文件, 数据/偏移量错误");
 
     init_psp_decrypt(&psp_key, psp_iv, 0, mac, key_header, 0x70, 0x30);
-
     static constexpr uint32_t block_size = 0x10;
     uint32_t block_count = ((data_size + (block_size - 1)) / block_size);
 
@@ -854,7 +852,8 @@ void Download::download_file_content_to_edat(uint64_t item_size)
     }
 
     if (!pkgi_write(item_file, data.data(), data.size()))
-        throw DownloadError(fmt::format("failed to write to %s", item_path));
+        throw DownloadError(fmt::format("无法写入至 %s", item_path));
+
     skip_to_file_offset(item_size);
 }
 
@@ -874,7 +873,7 @@ int Download::download_files(void)
     for (; item_index < index_count; ++item_index)
     {
         if (is_canceled())
-            throw std::runtime_error("download was canceled");
+            throw std::runtime_error("已取消下载");
 
         uint8_t item[32];
         pkgi_memcpy(
@@ -890,7 +889,7 @@ int Download::download_files(void)
         const uint8_t type = item[27];
 
         if (enc_offset + name_offset + name_size > total_size)
-            throw DownloadError("pkg file is too small or corrupted");
+            throw DownloadError("PKG文件不完整或已损坏");
 
         {
             std::vector<uint8_t> item_name_v(
@@ -982,20 +981,20 @@ int Download::download_files(void)
         {
             open_file();
             if (pkgi_seek(item_file, encrypted_offset) < 0)
-                throw ResumeError("failed to seek for resume");
+                throw ResumeError("无法恢复下载");
         }
         else
             create_file();
 
         if (enc_offset + item_offset + encrypted_offset != download_offset)
             throw formatEx<DownloadError>(
-                    "pkg is not supported, file offset mismatch, "
-                    "expected: {}, actual: {}",
+                    "PKG文件不受支持, 文件偏移值错误, "
+                    "预期: {}, 实际: {}",
                     enc_offset + item_offset + encrypted_offset,
                     +download_offset);
 
         if (enc_offset + item_offset + item_size > total_size)
-            throw DownloadError("pkg file is too small or corrupted");
+            throw DownloadError("PKG文件不完整或已损坏");
 
         if (content_type == CONTENT_TYPE_PSP_GAME ||
             content_type == CONTENT_TYPE_PSP_GAME_ALT ||
@@ -1088,7 +1087,7 @@ int Download::check_integrity(const uint8_t* digest)
 
         pkgi_rm(fmt::format("{}/sce_sys/package/head.bin", root).c_str());
 
-        throw DownloadError("pkg integrity failed, try downloading again");
+        throw DownloadError("PKG文件不完整或已损坏, 请尝试重新下载");
     }
 
     LOG("pkg integrity check succeeded");
@@ -1325,7 +1324,7 @@ void Download::deserialize_state()
         uint8_t version;
         iarchive(version);
         if (version != 1)
-            throw std::runtime_error("invalid resume data version");
+            throw std::runtime_error("无效的恢复数据版本");
 
         iarchive(save_as_iso);
         iarchive(download_offset, download_size);
@@ -1354,6 +1353,6 @@ void Download::deserialize_state()
     catch (const std::exception& e)
     {
         throw formatEx<ResumeError>(
-                "can't resume download:\n{}\nResume data deleted.", e.what());
+                "无法恢复下载:\n{}\n恢复数据已删除", e.what());
     }
 }