+++ src/install.cpp
@@ -85,10 +85,10 @@ void pkgi_install(const char* partition, const char* contentid)
     const auto res = scePromoterUtilityPromotePkgWithRif(path, 1);
     if (res < 0)
         throw formatEx<std::runtime_error>(
-                "scePromoterUtilityPromotePkgWithRif failed: {:#08x}\n{}",
+                "调用NoNpDrm函数错误: {:#08x}\n{}",
                 static_cast<uint32_t>(res),
                 static_cast<uint32_t>(res) == 0x80870004
-                        ? "Please check your NoNpDrm installation"
+                        ? "请检查NoNpDrm插件安装是否正确"
                         : "");
 }
 
@@ -103,7 +103,7 @@ void pkgi_install_update(const std::string& partition, const std::string& titlei
     const auto res = scePromoterUtilityPromotePkgWithRif(src.c_str(), 1);
     if (res < 0)
         throw formatEx<std::runtime_error>(
-                "scePromoterUtilityPromotePkgWithRif failed: {:#08x}",
+                "调用NoNpDrm函数错误: {:#08x}",
                 static_cast<uint32_t>(res));
 }
 
@@ -176,7 +176,7 @@ void pkgi_install_psmgame(const char* partition, const char* contentid)
     int res = sceIoRename(src.c_str(), dest.c_str());
     if (res < 0)
         throw formatEx<std::runtime_error>(
-                "failed to rename: {:#08x}", static_cast<uint32_t>(res));
+                "重命名失败: {:#08x}", static_cast<uint32_t>(res));
 
     LOGF("promoting psm game at {}", dest);
     ScePromoterUtilityImportParams promote_args;
@@ -190,7 +190,7 @@ void pkgi_install_psmgame(const char* partition, const char* contentid)
     res = scePromoterUtilityPromoteImport(&promote_args);
     if (res < 0)
         throw formatEx<std::runtime_error>(
-            "scePromoterUtilityPromoteImport failed: {:#08x}",
+            "scePromoterUtilityPromoteImport 失败: {:#08x}",
             static_cast<uint32_t>(res));
 }
 
@@ -207,7 +207,7 @@ void pkgi_install_pspgame(const char* partition, const char* gpath, const char* 
     int res = sceIoRename(path.c_str(), dest.c_str());
     if (res < 0)
         throw std::runtime_error(fmt::format(
-                "failed to rename: {:#08x}", static_cast<uint32_t>(res)));
+                "无法重命名: {:#08x}", static_cast<uint32_t>(res)));
 }
 
 static void pkgi_move_merge(const std::string& from, const std::string& to)
@@ -219,9 +219,9 @@ static void pkgi_move_merge(const std::string& from, const std::string& to)
         (fromType == InodeType::File && toType == InodeType::File))
         pkgi_rename(from, to);
     else if (fromType == InodeType::File && toType == InodeType::Directory)
-        throw formatEx("cannot replace directory {} by file {}", to, from);
+        throw formatEx("无法用{}文件替换{}文件夹", to, from);
     else if (fromType == InodeType::Directory && toType == InodeType::File)
-        throw formatEx("cannot replace directory {} by file {}", from, to);
+        throw formatEx("无法用{}文件替换{}文件夹", from, to);
     else if (fromType == InodeType::Directory && toType == InodeType::Directory)
     {
         const auto fromContents = pkgi_list_dir_contents(from);
@@ -232,7 +232,7 @@ static void pkgi_move_merge(const std::string& from, const std::string& to)
     }
     else
         throw formatEx(
-                "cannot merge {} ({}) and {} ({})",
+                "无法将 {} ({}) 和 {} ({})归并",
                 from,
                 (int)fromType,
                 to,